refactor: split the deferred-capture helpers out of fix_cycle
fix_cycle.py crossed the 1500-line first-party limit. The #305 defer capture
path (tracking-issue marker, conversation rendering, capture, and the
workflow-scope requeue) moves to fix_cycle_deferred; fix_cycle re-exports the
names its existing importers and monkeypatch sites already use. Behavior is
unchanged.
